Separate migrate and migration classes
